+++ src/main/java/org/organicdesign/fp/collections/MutSet.java
@@ -14,7 +14,9 @@
 
 package org.organicdesign.fp.collections;
 
+import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 /**
  Interface for mutable (hash) set builder.
@@ -30,13 +32,41 @@ public interface MutSet<E> extends BaseSet<E> {
     MutSet<E> put(E val);
 
     /** {@inheritDoc} */
-    @NotNull
-    @Override default MutSet<E> union(Iterable<? extends E> iter) {
-        return concat(iter).toMutSet();
+    @Override
+    @Contract(mutates = "this")
+    default @NotNull MutSet<E> union(Iterable<? extends E> iter) {
+        if (iter != null) {
+            for (E item : iter) {
+                put(item);
+            }
+        }
+        return this;
     }
 
     /** {@inheritDoc} */
     @NotNull
     @Override
     MutSet<E> without(E key);
+
+    /**
+     * Efficiently adds items to this MutSet.  Ordering is ignored.
+     * @param items the values to add
+     * @return a new MutSet with the additional items.
+     */
+    @Override
+    @Contract(mutates = "this")
+    default @NotNull MutSet<E> concat(@Nullable Iterable<? extends E> items) {
+        return union(items);
+    }
+
+    /**
+     * Efficiently adds items to this MutSet.  Ordering is ignored.
+     * @param items the values to add
+     * @return a new MutSet with the additional items.
+     */
+    @Override
+    @Contract(mutates = "this")
+    default @NotNull MutSet<E> precat(@Nullable Iterable<? extends E> items) {
+        return union(items);
+    }
 }